refactor: test 코드 패키지 위치 수정
