+++ src/main/resources/application-local.yml
@@ -5,7 +5,7 @@ spring:
 
 logging:
   level:
-    com.team3: debug
+    com.team3: info
     root: info
-    org.hibernate.SQL: debug
-    org.hibernate.orm.jdbc.bind: trace
+    org.hibernate.SQL: info
+    org.hibernate.orm.jdbc.bind: info
